tests/requirements.txt
dotmap>=1.3.30
GitPython>=3.1.31
giturlparse>=0.10.0
pytest>=7.2.2
python_gitlab>=3.14.0
